+++ Form/UserDeleteRoleType.php
@@ -9,34 +9,45 @@
 
 namespace c975L\UserBundle\Form;
 
-use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * UserResetPassword FormType
+ * UserDeleteRoleType FormType
  * @author Laurent Marquet <laurent.marquet@laposte.net>
  * @copyright 2018 975L <contact@975l.com>
  */
-class UserResetPasswordType extends AbstractType
+class UserDeleteRoleType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        $roles = $options['config']['user']->getRoles();
+        $roles = array_combine($roles, $roles);
+        unset($roles['ROLE_USER']);
+
         $builder
-            ->add('email', EmailType::class, array(
-                'label' => 'label.email',
+            ->add('user', TextType::class, array(
+                'label' => 'label.user',
                 'required' => true,
+                'mapped' => false,
+                'data' => $options['config']['user']->getEmail(),
                 'attr' => array(
-                    'placeholder' => 'placeholder.email',
+                    'readonly' => true,
                 )))
+            ->add('role', ChoiceType::class, array(
+                'mapped' => false,
+                'choices' => $roles,
+                ))
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults(array(
-            'intention' => 'UserResetPasswordForm',
+            'intention' => 'UserDeleteRoleForm',
             'translation_domain' => 'user',
         ));
