+++ Resources/config/services.yml
@@ -15,6 +15,8 @@ services:
 
     c975L\UserBundle\Security\ApiVoter:
         public: true
+    c975L\UserBundle\Security\ManageVoter:
+        public: true
     c975L\UserBundle\Security\UserVoter:
         public: true
     c975L\UserBundle\Security\TokenAuthenticator:
