+++ Resources/views/pages/users.html.twig
@@ -1,0 +1,48 @@
+{% extends "@c975LUser/layout.html.twig" %}
+{% trans_default_domain 'user' %}
+{% set title = 'label.users'|trans({}, 'toolbar') %}
+
+{% block user_content %}
+{# Toolbar #}
+    {{ toolbar_display('@c975LUser/tools.html.twig', 'users', 'md', currentUser) }}
+{# List of users and actions #}
+    <div class="text-center">
+        <h2>
+            {{ 'label.users'|trans({}, 'toolbar') }}
+        </h2>
+        <table class="table table-hover table-condensed table-responsive">
+            <thead>
+                <tr>
+                    <th class="text-center"><span class="fas fa-eye"></span>&nbsp;{{ knp_pagination_sortable(users, 'label.email'|trans, 'email') }}&nbsp;<span class="fas fa-sort"></span></th>
+                    <th class="text-center">{{ 'label.manage'|trans({}, 'toolbar') }}</th>
+                </tr>
+            </thead>
+            <tbody>
+                {% for user in users %}
+                    <tr>
+{# Email #}
+                        <td class="text-center">
+                            {{ user.email }}
+                        </td>
+                        <td class="text-center">
+{# Display #}
+                            {{ toolbar_button(path('user_manage_display', {'identifier': user.identifier}), 'display', 'sm') }}
+{# Modify #}
+                            {{ toolbar_button(path('user_manage_modify', {'identifier': user.identifier}), 'modify', 'sm') }}
+{# Add Role #}
+                            {{ toolbar_button(path('user_manage_add_role', {'identifier': user.identifier}), 'add_role', 'sm') }}
+{# Delete Role #}
+                            {{ toolbar_button(path('user_manage_delete_role', {'identifier': user.identifier}), 'delete_role', 'sm') }}
+{# Delete #}
+                            {{ toolbar_button(path('user_manage_delete', {'identifier': user.identifier}), 'delete', 'sm') }}
+                        </td>
+                    </tr>
+                {% endfor %}
+            </tbody>
+        </table>
+{# Pagination #}
+        <div class="navigation text-center">
+            {{ knp_pagination_render(users) }}
+        </div>
+    </div>
+{% endblock %}
