+++ Resources/views/tools.html.twig
@@ -24,8 +24,13 @@
     {{ toolbar_button(path('user_change_password'), 'change_password') }}
 {% endif %}
 
+{# Manage #}
+{% if type != 'users' and is_granted(config('c975LUser.roleNeeded')) %}
+    {{ toolbar_button(path('user_manage'), 'users') }}
+{% endif %}
+
 {# Config #}
-{% if type != 'config' and is_granted('ROLE_ADMIN') %}
+{% if type != 'config' and is_granted(config('c975LUser.roleNeeded')) %}
     {{ toolbar_button(path('user_config'), 'config') }}
 {% endif %}
